Remove unneeeded casts.
Reviewed-on: https://git.eclipse.org/r/7472
IP-Clean: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>

+++ gcov/org.eclipse.linuxtools.gcov.core/src/org/eclipse/linuxtools/internal/gcov/parser/GcdaRecordsParser.java
@@ -84,7 +84,7 @@ public class GcdaRecordsParser {
 				int tag = stream.readInt();
 				long length = (stream.readInt() &  MasksGenerator.UNSIGNED_INT_MASK);
 				// parse gcda data
-				switch ((int) tag) {
+				switch (tag) {
 				case GCOV_TAG_FUNCTION: {
 					long fnctnId = stream.readInt() &  MasksGenerator.UNSIGNED_INT_MASK;
 					if (!fnctns.isEmpty()) {
